+++ tests/unit/test_metadata.py
@@ -15,7 +15,7 @@ def test_cargo_env() -> None:
 @pytest.mark.metadata
 @pytest.mark.local
 @pytest.mark.parametrize(
-    "backend", ("poetry", "uv", None)
+    "backend", ("poetry", "uv", "conda", None)
 )
 def test_python_env(backend: str | None) -> None:
     if backend == "poetry":
@@ -24,6 +24,9 @@ def test_python_env(backend: str | None) -> None:
     elif backend == "uv":
         metadata = sv_meta._python_env(pathlib.Path(__file__).parents[1].joinpath("example_data", "python_uv"))
         assert metadata["project"]["name"] == "example-repo"
+    elif backend == "conda":
+        metadata = sv_meta._python_env(pathlib.Path(__file__).parents[1].joinpath("example_data", "python_conda"))
+        assert metadata["environment"]["requests"]
     else:
         metadata = sv_meta._python_env(pathlib.Path(__file__).parents[1].joinpath("example_data"))
 
@@ -54,7 +57,6 @@ def test_environment() -> None:
     assert metadata["julia"]["project"]["name"] == "Julia Demo Project"
     assert metadata["javascript"]["project"]["name"] == "my-awesome-project"
 
-
 @pytest.mark.metadata
 @pytest.mark.local
 def test_slurm_env_var_capture() -> None:
@@ -75,3 +77,4 @@ def test_slurm_env_var_capture() -> None:
 
     sv_meta.metadata = sv_meta.environment(env_var_glob_exprs={"SLURM_*"})
     assert all((key, value) in sv_meta.metadata["shell"].items() for key, value in _slurm_env.items())
+